Refine legacy handler return inference
