+++ src/BlingoEngine.Legacy.Lingo/Analysis/Blocks/IHandlerBlockClassifier.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+using BlingoEngine.Legacy.Lingo.Syntax;
+
+namespace BlingoEngine.Legacy.Lingo.Analysis.Blocks;
+
+internal interface IHandlerBlockClassifier
+{
+    bool TryCreate(IReadOnlyList<BlSyntaxToken> tokens, BlLingoSymbolTable symbols, out BlLingoHandlerCodeBlock block);
+}
