+++ src/BlingoEngine.Legacy.Lingo/CodeGen/BlLegacyHandlerConverter.cs
@@ -16,6 +16,7 @@ public sealed class BlLegacyHandlerConverter
     private readonly BlLegacyClassGeneratorOptions _options;
     private readonly BlLingoAnalysisResult _analysis;
     private readonly IReadOnlyDictionary<BlLingoHandlerSymbolTable, IReadOnlyList<BlLingoHandlerCodeBlock>> _handlerBlocks;
+    private readonly IReadOnlyDictionary<BlLingoHandlerSymbolTable, string?> _handlerReturnTypes;
 
     public BlLegacyHandlerConverter(
         string source,
@@ -30,13 +31,21 @@ public sealed class BlLegacyHandlerConverter
 
         IReadOnlyDictionary<BlLingoHandlerSymbolTable, IReadOnlyList<BlLingoHandlerCodeBlock>> blocks =
             new Dictionary<BlLingoHandlerSymbolTable, IReadOnlyList<BlLingoHandlerCodeBlock>>();
+        IReadOnlyDictionary<BlLingoHandlerSymbolTable, string?> returnTypes =
+            new Dictionary<BlLingoHandlerSymbolTable, string?>();
 
         if (analysis.TryGetData(BlLingoHandlerCodeBlockPass.HandlerCodeBlocksKey, out IReadOnlyDictionary<BlLingoHandlerSymbolTable, IReadOnlyList<BlLingoHandlerCodeBlock>>? extracted) && extracted is not null)
         {
             blocks = extracted;
         }
 
+        if (analysis.TryGetData(BlLingoHandlerCodeBlockPass.HandlerReturnTypesKey, out IReadOnlyDictionary<BlLingoHandlerSymbolTable, string?>? inferredReturnTypes) && inferredReturnTypes is not null)
+        {
+            returnTypes = inferredReturnTypes;
+        }
+
         _handlerBlocks = blocks;
+        _handlerReturnTypes = returnTypes;
     }
 
     /// <summary>
@@ -63,6 +72,70 @@ public sealed class BlLegacyHandlerConverter
         var fallbackBlocks = BlLingoHandlerFallbackBuilder.Build(body.Tokens, body.EndLeadingTrivia, _analysis.Symbols);
         var fallbackEmitter = new BlLegacyHandlerBodyEmitter(writer, fallbackBlocks, _options, _analysis);
         fallbackEmitter.Emit();
+    }
+
+    public string? GetReturnType(BlLingoHandlerSymbolTable handler)
+    {
+        if (handler is null)
+        {
+            return null;
+        }
+
+        if (_handlerReturnTypes.TryGetValue(handler, out var inferred) && !string.IsNullOrWhiteSpace(inferred))
+        {
+            return inferred;
+        }
+
+        return null;
+    }
+
+    public string? InferReturnType(BlLingoHandlerSymbolTable handler)
+    {
+        if (handler is null)
+        {
+            return null;
+        }
+
+        if (!_handlerBlocks.TryGetValue(handler, out var blocks) || blocks.Count == 0)
+        {
+            return null;
+        }
+
+        string? inferred = null;
+        foreach (var block in blocks)
+        {
+            var candidate = BlLegacyReturnTypeHelper.InferBlockResult(block);
+            if (!string.IsNullOrEmpty(candidate))
+            {
+                inferred = BlLegacyReturnTypeHelper.Merge(inferred, candidate);
+            }
+        }
+
+        return inferred;
+    }
+
+    public bool HasReturnValue(BlLingoHandlerSymbolTable handler)
+    {
+        if (handler is null)
+        {
+            return false;
+        }
+
+        if (_handlerBlocks.TryGetValue(handler, out var blocks))
+        {
+            foreach (var block in blocks)
+            {
+                if (block.Kind == BlLingoHandlerCodeBlockKind.Expression && block.Data is BlLingoExpressionBlockData expressionData)
+                {
+                    if (BlLegacyReturnTypeHelper.IsReturnWithValue(expressionData.Expression))
+                    {
+                        return true;
+                    }
+                }
+            }
+        }
+
+        return false;
     }
 
     private HandlerBody ExtractHandlerBody(BlLingoHandlerSymbolTable handler)
